Added option for NUM_PLACEHOLDER to represent item number
Added option for the NUM_PLACEHOLDER to represent either page number or item number. This may be useful for situations when the number in the url parameters represents item number instead of page number, which is typical for some applications, such as Apache Solr (see for example the start parameter described at https://cwiki.apache.org/confluence/x/YoDxAQ ). The item number is then set to the number of the first item on the page and the item numbers start with zero.